@@ -16,49 +16,50 @@ using Xunit.Abstractions;
 
 namespace MediaInfo.Wrapper.Tests
 {
-  public class TestLogger : ILogger
-  {
-    private readonly ITestOutputHelper _testOutputHelper;
+    public class TestLogger : ILogger
+    {
+        private readonly ITestOutputHelper _testOutputHelper;
 #if !NET5_0_OR_GREATER
-    private readonly Regex _regex = new(@"\{(?<logValue>[^\}]+)\}", RegexOptions.Singleline | RegexOptions.Compiled);
+        private readonly Regex _regex = new(@"\{(?<logValue>[^\}]+)\}", RegexOptions.Singleline | RegexOptions.Compiled);
 #endif
 
         public TestLogger(ITestOutputHelper testOutputHelper)
-    {
-      _testOutputHelper = testOutputHelper;
-    }
+        {
+            _testOutputHelper = testOutputHelper;
+        }
 
 #if NET5_0_OR_GREATER
-    public IDisposable BeginScope<TState>(TState state)
+    public IDisposable? BeginScope<TState>(TState state) where TState : notnull
     {
         throw new NotImplementedException();
     }
 
     public bool IsEnabled(LogLevel logLevel) => true;
 
-    public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
+    public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception? exception, Func<TState, Exception?, string> formatter)
     {
       _testOutputHelper.WriteLine($"{logLevel}: {formatter(state, exception)}");
     }
 #else
-    public void Log(LogLevel loglevel, string message, params object[] parameters)
-    {
-      var processedMessage = message;
-      var position = 0;
-      var index = 0;
-      foreach (var parameter in parameters)
-      {
-        var result = _regex.Match(processedMessage, position);
-        if (result.Success)
+        public void Log(LogLevel loglevel, string message, params object?[] parameters)
         {
-          processedMessage = processedMessage.Replace(result.Value, $"{{{index}}}");
-          position = result.Index + 1;
-        }
+            var processedMessage = message;
+            var position = 0;
+            var index = 0;
+            foreach (var parameter in parameters)
+            {
+                var result = _regex.Match(processedMessage, position);
+                if (result.Success)
+                {
+                    processedMessage = processedMessage.Replace(result.Value, $"{{{index}}}");
+                    position = result.Index + 1;
+                }
 
-        index++;
-      }
-      _testOutputHelper.WriteLine($"{loglevel}: {string.Format(processedMessage, parameters)}");
-    }
+                index++;
+            }
+
+            _testOutputHelper.WriteLine($"{loglevel}: {string.Format(processedMessage, parameters)}");
+        }
 #endif
-  }
+    }
 }
