@@ -1,18 +1,25 @@
-﻿namespace MediaInfo.Model
+﻿#region Copyright (C) 2017-2022 Yaroslav Tatarenko
+
+// Copyright (C) 2017-2022 Yaroslav Tatarenko
+// This product uses MediaInfo library, Copyright (c) 2002-2020 MediaArea.net SARL. 
+// https://mediaarea.net
+
+#endregion
+
+namespace MediaInfo.Model;
+
+/// <summary>
+/// Describes frame rate mode
+/// </summary>
+public enum FrameRateMode
 {
     /// <summary>
-    /// Describes frame rate mode
+    /// Frame rate is constant
     /// </summary>
-    public enum FrameRateMode
-    {
-        /// <summary>
-        /// Frame rate is constant
-        /// </summary>
-        Constant,
+    Constant,
 
-        /// <summary>
-        /// Frame rate varies
-        /// </summary>
-        Variable
-    }
+    /// <summary>
+    /// Frame rate varies
+    /// </summary>
+    Variable
 }
