@@ -25,7 +25,7 @@ public class AudioTags : BaseTags
     /// <example>The title of the album</example>
     [DataMember(Name = "album")]
     [JsonPropertyName("album")]
-    public string Album { get; set; }
+    public string? Album { get; set; }
 
     /// <summary>
     /// The title of the track.
@@ -33,7 +33,7 @@ public class AudioTags : BaseTags
     /// <example>The title of the track</example>
     [DataMember(Name = "track")]
     [JsonPropertyName("track")]
-    public string Track { get; set; }
+    public string? Track { get; set; }
 
     /// <summary>
     /// The title of the subtrack.
@@ -41,7 +41,7 @@ public class AudioTags : BaseTags
     /// <example>The title of the subtrack</example>
     [DataMember(Name = "subTrack")]
     [JsonPropertyName("subTrack")]
-    public string SubTrack { get; set; }
+    public string? SubTrack { get; set; }
 
     /// <summary>
     /// The original album name (in case of a remake/remix).
@@ -49,7 +49,7 @@ public class AudioTags : BaseTags
     /// <example>The title</example>
     [DataMember(Name = "originalAlbum")]
     [JsonPropertyName("originalAlbum")]
-    public string OriginalAlbum { get; set; }
+    public string? OriginalAlbum { get; set; }
 
     /// <summary>
     /// Gets the original track name (in case of a remake/remix).
@@ -57,7 +57,7 @@ public class AudioTags : BaseTags
     /// <example>The title of the track</example>
     [DataMember(Name = "originalTrack")]
     [JsonPropertyName("originalTrack")]
-    public string OriginalTrack { get; set; }
+    public string? OriginalTrack { get; set; }
 
     /// <summary>
     /// The number of the current track.
@@ -97,7 +97,7 @@ public class AudioTags : BaseTags
     /// <example>Bob Dylan</example>
     [DataMember(Name = "artist")]
     [JsonPropertyName("artist")]
-    public string Artist { get; set; }
+    public string? Artist { get; set; }
 
     /// <summary>
     /// An album artist.
@@ -105,7 +105,7 @@ public class AudioTags : BaseTags
     /// <example>Bob Dylan</example>
     [DataMember(Name = "albumArtist")]
     [JsonPropertyName("albumArtist")]
-    public string AlbumArtist { get; set; }
+    public string? AlbumArtist { get; set; }
 
     /// <summary>
     /// An official artist/performer web page.
@@ -113,14 +113,14 @@ public class AudioTags : BaseTags
     /// <example>http://www.bobdylan.com</example>
     [DataMember(Name = "artistUrl")]
     [JsonPropertyName("artistUrl")]
-    public string ArtistUrl { get; set; }
+    public string? ArtistUrl { get; set; }
 
     /// <summary>
     /// A accompaniment name.
     /// </summary>
     [DataMember(Name = "accompaniment")]
     [JsonPropertyName("accompaniment")]
-    public string Accompaniment { get; set; }
+    public string? Accompaniment { get; set; }
 
     /// <summary>
     /// A composer name.
@@ -128,7 +128,7 @@ public class AudioTags : BaseTags
     /// <example>Bob Dylan</example>
     [DataMember(Name = "composer")]
     [JsonPropertyName("composer")]
-    public string Composer { get; set; }
+    public string? Composer { get; set; }
 
     /// <summary>
     /// A composer nationality.
@@ -136,14 +136,14 @@ public class AudioTags : BaseTags
     /// <example>USA</example>
     [DataMember(Name = "composerNationality")]
     [JsonPropertyName("composerNationality")]
-    public string ComposerNationality { get; set; }
+    public string? ComposerNationality { get; set; }
 
     /// <summary>
     /// An arranger name.
     /// </summary>
     [DataMember(Name = "arranger")]
     [JsonPropertyName("arranger")]
-    public string Arranger { get; set; }
+    public string? Arranger { get; set; }
 
     /// <summary>
     /// A lyricist name.
@@ -151,42 +151,42 @@ public class AudioTags : BaseTags
     /// <example>Bob Dylan</example>
     [DataMember(Name = "lyricist")]
     [JsonPropertyName("lyricist")]
-    public string Lyricist { get; set; }
+    public string? Lyricist { get; set; }
 
     /// <summary>
     /// A conductor name.
     /// </summary>
     [DataMember(Name = "conductor")]
     [JsonPropertyName("conductor")]
-    public string Conductor { get; set; }
+    public string? Conductor { get; set; }
 
     /// <summary>
     /// A sound engineer name.
     /// </summary>
     [DataMember(Name = "soundEngineer")]
     [JsonPropertyName("soundEngineer")]
-    public string SoundEngineer { get; set; }
+    public string? SoundEngineer { get; set; }
 
     /// <summary>
     /// Who mastered track.
     /// </summary>
     [DataMember(Name = "masteredBy")]
     [JsonPropertyName("masteredBy")]
-    public string MasteredBy { get; set; }
+    public string? MasteredBy { get; set; }
 
     /// <summary>
     /// Who remixed track.
     /// </summary>
     [DataMember(Name = "remixedBy")]
     [JsonPropertyName("remixedBy")]
-    public string RemixedBy { get; set; }
+    public string? RemixedBy { get; set; }
 
     /// <summary>
     /// A label name.
     /// </summary>
     [DataMember(Name = "label")]
     [JsonPropertyName("label")]
-    public string Label { get; set; }
+    public string? Label { get; set; }
 
     /// <summary>
     /// The recorded date.
@@ -202,54 +202,54 @@ public class AudioTags : BaseTags
     /// <example>rock</example>
     [DataMember(Name = "genre")]
     [JsonPropertyName("genre")]
-    public string Genre { get; set; }
+    public string? Genre { get; set; }
 
     /// <summary>
     /// The mood.
     /// </summary>
     [DataMember(Name = "mood")]
     [JsonPropertyName("mood")]
-    public string Mood { get; set; }
+    public string? Mood { get; set; }
 
     /// <summary>
     /// The track ISRC.
     /// </summary>
     [DataMember(Name = "isrc")]
     [JsonPropertyName("isrc")]
-    public string Isrc { get; set; }
+    public string? Isrc { get; set; }
 
     /// <summary>
     /// The bar code.
     /// </summary>
     [DataMember(Name = "barCode")]
     [JsonPropertyName("barCode")]
-    public string BarCode { get; set; }
+    public string? BarCode { get; set; }
 
     /// <summary>
     /// The LCCN.
     /// </summary>
     [DataMember(Name = "lccn")]
     [JsonPropertyName("lccn")]
-    public string Lccn { get; set; }
+    public string? Lccn { get; set; }
 
     /// <summary>
     /// The catalog number.
     /// </summary>
     [DataMember(Name = "catalogNumber")]
     [JsonPropertyName("catalogNumber")]
-    public string CatalogNumber { get; set; }
+    public string? CatalogNumber { get; set; }
 
     /// <summary>
     /// The label code.
     /// </summary>
     [DataMember(Name = "labelCode")]
     [JsonPropertyName("labelCode")]
-    public string LabelCode { get; set; }
+    public string? LabelCode { get; set; }
 
     /// <summary>
     /// The name of the person or organization that encoded/ripped the audio file.
     /// </summary>
     [DataMember(Name = "encodedBy")]
     [JsonPropertyName("encodedBy")]
-    public string EncodedBy { get; set; }
+    public string? EncodedBy { get; set; }
 }
